+++ vNES/src/vnes/Mapper004.java
@@ -66,7 +66,7 @@ public class Mapper004 extends MapperDefault {
         super.mapperInternalStateSave(buf);
 
         // Version:
-        buf.putByte((int) 1);
+        buf.putByte( 1);
 
         // State:
         buf.putInt(command);
