+++ vNES/src/vnes/Mapper009.java
@@ -189,7 +189,7 @@ public class Mapper009 extends MapperDefault {
         super.mapperInternalStateSave(buf);
 
         // Version:
-        buf.putByte((int) 1);
+        buf.putByte( 1);
 
         // State:
         buf.putByte((byte) latchLo);
